resetting this branch to the working set of files from staging.
At this point in the project things were in two separate repos. This branch looks to have been initially used to bring in the Dayton files, but then quickly forgotten. Since there wasn't a single branch in either this or the responsive_example_set repo that represented the Truman specific responsiveness, I decided to repurpose this branch based purely on name.